104-fibonacci.c

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int i = 0;
-	long j = 1, k = 2;
+	int j = 1, k = 2;
 
 	while (i < 98)
 	{
